search/pyproject.toml
[project]
name = "documents-api"
version = "2.0.0"
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
  "fastapi[standard]==0.116.2",
  "sqlmodel>=0.0.24,<0.1.0",
  "pydantic-settings>=2.9.1,<3.0.0",
  "psycopg2>=2.9.10,<3.0.0",
  "pyvespa>=0.61.0",
]

[dependency-groups]
dev = ["ty>=0.0.1a20"]
